+++ extension/src/utils.ts
@@ -1,17 +1,31 @@
-import { window, Selection } from 'vscode';
-import { defaultComponentName } from './constants';
+import { Selection, window } from 'vscode';
+import { defaultComponentName, defaultSkipImport } from './constants';
 
 export async function showInputBox(): Promise<string> {
-    const result = await window.showInputBox({
-        value: '',
-        placeHolder: 'Choose your component name ...',
-    });
-    return result || defaultComponentName;
+  const result = await window.showInputBox({
+    value: '',
+    placeHolder: 'Choose your component name ...',
+  });
+  return result || defaultComponentName;
 }
 
-export function createScript(newComponent: string, activeFile: string, selection: Selection): string {
-    const { start, end } = selection;
-    const script = `ng g @componizer/schematics:ng-componize --name ${newComponent} --activeFile ${activeFile} --start ${start.line} --end ${end.line}`;
-    window.showInformationMessage(script);
-    return script;
+export async function showInputBoxForSkipImport(): Promise<string> {
+  const result = await window.showInputBox({
+    value: '',
+    placeHolder: 'Allows for skipping the module import.',
+  });
+  return result || defaultSkipImport;
+}
+
+export function createScript(
+  newComponent: string,
+  skipImport: string,
+  activeFile: string,
+  selection: Selection,
+  debugMode: boolean
+): string {
+  const { start, end } = selection;
+  const script = `ng g @componizer/schematics:ng-componize --name ${newComponent} --activeFile ${activeFile} --start ${start.line} --end ${end.line} --customSkipImport ${skipImport} --debugMode ${debugMode}`;
+  window.showInformationMessage(script);
+  return script;
 }